## Solution
The developer changed the version of the package from 0.0.1 to 0.0.2.

## Explanation

The diff file is a standard format for showing the differences between two files.

The first line shows the file that was changed.

The second line shows the index of the file.

The third line shows the old version of the file.

The fourth line shows the new version of the file.

The lines that start with a + are the lines that were added.

The lines that start with a - are the lines that were removed.

The lines that start with a @@ are the lines that were changed.

The lines that start with a number are the lines that were not changed.

The lines that start with a space are the lines that were not changed.

The lines that start with a - are the lines that were removed.

The lines that start with a + are the lines that were added.

The lines that start with a @@ are the lines that were changed.

The lines that start with a number are the lines that were not changed.